requirements.txt
Flask==2.3.5
google-generativeai

